# CodeGen: Improve lowering of NUM_TO_VEC on A64 for constants