--- a/tests/conformance/vector.lua
+++ b/tests/conformance/vector.lua
@@ -51,6 +51,9 @@ assert(8 * vector(8, 16, 24) == vector(64, 128, 192));
 assert(vector(1, 2, 4) * '8' == vector(8, 16, 32));
 assert('8' * vector(8, 16, 24) == vector(64, 128, 192));
 
+assert(vector(1, 2, 4) * 100 == vector(100, 200, 400))
+assert(100 * vector(1, 2, 4) == vector(100, 200, 400))
+
 if vector_size == 4 then
 	assert(vector(1, 2, 4, 8) / vector(8, 16, 24, 32) == vector(1/8, 2/16, 4/24, 8/32));
 	assert(8 / vector(8, 16, 24, 32) == vector(1, 1/2, 1/3, 1/4));